TASK-1/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
faiss-cpu
sentence-transformers
groq
requests
nltk
numpy
pytest
pytest-asyncio
hypothesis
python-dotenv
pypdf